adding checker for repo route within the fast api and some refactoring of frontend components
