+++ webui/src/types/API.d.ts
@@ -1,0 +1,4 @@
+export type Status = {
+  id: number;
+  status: number;
+};
